feat: refactor MoviesList to use MovieCard component for better code organization
