+++ src/components/movies/MoviesList.tsx
@@ -1,42 +1,19 @@
 // displays all movies in cart view in homepage
-import { useNavigate } from "react-router-dom";
 import { TrendingMoviesData } from "../../types/Index";
+import { MovieCard } from "./MovieCard";
 
 type MovieListProps = {
   popularMovies: TrendingMoviesData[];
 };
 
 const MoviesList = ({ popularMovies }: MovieListProps) => {
-  const navigate = useNavigate();
-
-  //handle movie Click in homepage
-  const handleMovieClick = (movieId: number) => {
-    navigate(`/movie/${movieId}`);
-  };
-
   return (
     <div className="movies">
       <h1 className="text-center m-5 title">Trending Movies</h1>
       <div className="row d-flex g-5">
         {popularMovies.map((movie, id) => (
           <div className="col-lg-4 mh-25" key={id}>
-            <div
-              className="card p-3"
-              onClick={() => handleMovieClick(movie.id)}
-            >
-              <h3 className="text-center">{movie.title || movie.name}</h3>
-              <div className="movie-list">
-                <img
-                  src={`https://image.tmdb.org/t/p/w400${movie.poster_path}`}
-                  alt={movie.title}
-                  className="img-fluid"
-                />
-                <p className="mt-4">
-                  <span className="bold">Plot: </span>
-                  {movie.overview}
-                </p>
-              </div>
-            </div>
+            <MovieCard movie={movie} />
           </div>
         ))}
       </div>
